controller08102021-2

+++ src/main/java/br/gov/sp/fatec/springlojaapp/service/ProdutoService.java
@@ -14,7 +14,7 @@ public interface ProdutoService {
 	
 	public Produto pesquisarPorNomeProduto (String nome);
 	
-	public Produto pesquisarPorIdProduto (int i);
+	public Produto pesquisarPorIdProduto (Long i);
 	
 	public List<Produto> pesquisarTodosProdutos ();
